Ll/event restructuring (#70)
This is a relatively small pr which blew up since I had to regenerate
the cassettes.

The goal if this pr is to unify our event rules. There was weird
behavior in the ui because we are doing different things in different
scenarios.

Specifically: It should be impossible to end in anything other than an
end event (be it success, error, or cancel)

Now we are done once we get our first end event. Anybody who needs to
drop something in (ie state change), must do it when they receive this
end event.

This simplifies our logic since nobody needs to hold onto events to
defer then anymore. Simply continue iterating and ignore future events
(while logging warnings)