open telemetry (#181)
In this PR

1. Implement our fastapi app with opentelemetry
2. Propagate opentelemetry headers so that spans are tracked throughout
the system
3. Add a few spans on important portions of the code (ie llm request,
tool request, etc)
4. Added a quick readme so we can set it up easier

Issues: We cannot create separate spans for each tool call. There is an
issue with merging streams with different spans. I have no idea how to
fix this and have spent a long time debugging it. I came up with a super
minimal repro which shows the issue in isolation outside of our
components, so I think that this is an otel bug (or there is some super
easy way around it and I just need to sleep.)

But it looks pretty good!