Ll/usage servie (#188)
Introduce two new projects, usage-service and usage-client. Separated so
that we can use the usage client in eidolon directly. Client also
contain the BaseModels that define the FastAPI definition.

The api includes the ability to modify the current used count by a delta
or set it to a specific number (reset being a subset of the last
category)

Usage service is NOT responsible for blocking requests, only observing
how users track to their current limits

Mongo queries will be efficient as long as there is a unique index on
"i". The service checks and adds the index on startup if it is missing

Approach: there is one mongo collection that records transactions. When
inserting a new transaction it gets the last transaction and tries to
insert the next transaction. If there is a race condition and another
transaction is inserted, it will error, catch the error, and tries
again. Retries are up to a limit (after which the http response is 503
indicating server too busy). Current usage + limit is recorded per
record, so getting current limits is a read to the latest (by field "i")
record.

Testing: there are tests along the service boundary that require mongo
to run. These tests also use the client directly to guarantee the client
functions properly.