Adds context to OpenAI not found errors in case of un/under-funded accounts
When an OpenAI token is used that either does not have billing enabled or does not
have enough credits to call the default model, their API will throw an exception
of type NotFoundError.

When this happens, we will catch the errors of this type and modify the message to
include additional context that indicates that the issue may be related to
insufficient credits. We will then re-raise an `APUException`.